% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAEI.R
\name{getAEI}
\alias{getAEI}
\title{Compute a design's augmented expected improvement (AEI)}
\usage{
getAEI(point, model, designs, ebs, se.threshold = 1e-06)
}
\arguments{
\item{point}{Numeric vector of same dimension as \code{model} and \code{designs}}

\item{model}{Object of class \code{km} with attr(,"package") = "DiceKriging"}

\item{designs}{Matrix with columns corresponding to design parameters}

\item{ebs}{Output of internal function \code{getEBS}}

\item{se.threshold}{Setting AEI to zero for nearby tested values}
}
\description{
\code{getAEI} returns the AEI for a proposed design points based
on a provided Gaussian Process model, the previously tested
design points, and the effective best solution (EBS) thus far.
}
\details{
TODO: PROVIDE DETAILS ON AEI COMPUTATION
}
